requirements.txt
beautifulsoup4
pandas
selenium
requests
python-docx
webdriver-manager
streamlit
openpyxl